# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Street Mech APIs",
        "description": "Street Mech helps drivers quickly find and order the nearest mechanic or tire repair service during car malfunctions in unfamiliar areas.",
        "version": "2.0.0"
    },
    "paths": {
...
        "/search": {
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "Search for shops",
                "description": "Searches for available shops in a specified city",
                "operationId": "289e44783351a6847216358a42a09a01",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "city"
                                ],
                                "properties": {
                                    "city": {
                                        "type": "string",
                                        "example": "beirut"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of shops found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "we have shops"
                                        },
                                        "state": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "shops": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "internal server error (problem in session)"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/order": {
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "Place an order",
                "description": "Creates an order for a car repair based on user and session data",
                "operationId": "ecfdac45a05fa0b861255b3aa9f05607",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "request is set"
                                        },
                                        "plate": {
                                            "type": "string",
                                            "example": "ABC123"
                                        },
                                        "brand": {
                                            "type": "string",
                                            "example": "Toyota"
                                        },
                                        "year_model": {
                                            "type": "string",
                                            "example": "2015"
                                        },
                                        "color": {
                                            "type": "string",
                                            "example": "Red"
                                        },
                                        "workerP": {
                                            "type": "string",
                                            "example": "1234567890"
                                        },
                                        "city": {
                                            "type": "string",
                                            "example": "beirut"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "internal server error (problem in session)"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format **Bearer &lt;token&gt;**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}